% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iter.accumulate}
\alias{iter.accumulate}
\title{itertools.accumulate}
\usage{
iter.accumulate(p, func = NULL)
}
\arguments{
\item{p}{}

\item{func}{}
}
\value{
list or vector
}
\description{
itertools.accumulate
}
\examples{
iter.accumulate(c(1, 2, 3, 4, 5))
iter.accumulate(c('A', 'B', 'C'))
iter.accumulate(c(1, 2, 3, 4, 5), func = function(x, y) { x * y })
}

