% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iter.compress}
\alias{iter.compress}
\title{itertools.compress}
\usage{
iter.compress(data, selectors)
}
\arguments{
\item{data}{iterables}

\item{selectors}{iterables}
}
\value{
list or vector
}
\description{
itertools.compress
}
\examples{
iter.compress(c(1, 2, 3), c(TRUE, FALSE, TRUE))
}

