% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iter.dropwhile}
\alias{iter.dropwhile}
\title{itertools.dropwhile}
\usage{
iter.dropwhile(pred, seq)
}
\arguments{
\item{pred}{iterables}

\item{seq}{iterables}
}
\value{
list or vector
}
\description{
itertools.dropwhile
}
\examples{
iter.dropwhile(function(x) { x > 2 }, c(1, 2, 3, 2, 1))
}

