% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iter.filterfalse}
\alias{iter.filterfalse}
\title{itertools.filterfalse}
\usage{
iter.filterfalse(pred, seq)
}
\arguments{
\item{pred}{iterables}

\item{seq}{iterables}
}
\value{
list or vector
}
\description{
itertools.filterfalse
}
\examples{
iter.filterfalse(function(x) { x > 2 }, c(1, 2, 3, 2, 1))
}

