% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iter.groupby}
\alias{iter.groupby}
\title{itertools.groupby}
\usage{
iter.groupby(...)
}
\arguments{
\item{iterable}{iterables}

\item{key}{}
}
\value{
list or vector
}
\description{
itertools.groupby
}
\examples{
iter.chain(c(1, 2, 3), c(1, 2, 3))
}

