% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iter.islice}
\alias{iter.islice}
\title{itertools.islice}
\usage{
iter.islice(iterable, ...)
}
\arguments{
\item{iterable}{iterables}

\item{start}{}

\item{stop}{}

\item{step}{}
}
\value{
list or vector
}
\description{
itertools.islice
}
\examples{
iter.islice('ABCDEFG', 2)
iter.islice('ABCDEFG', 2, 4)
iter.islice('ABCDEFG', 2, NULL)
iter.islice('ABCDEFG', 0, NULL, 2)
}

