% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iter.starmap}
\alias{iter.starmap}
\title{itertools.starmap}
\usage{
iter.starmap(func, seq)
}
\arguments{
\item{func}{}

\item{seq}{iterables}
}
\value{
list or vector
}
\description{
itertools.starmap
}
\examples{
iter.starmap(function(x, y, z) { x + y + z }, list(c(1, 2, 3), c(2, 3, 4)))
}

