% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iter.takewhile}
\alias{iter.takewhile}
\title{itertools.takewhile}
\usage{
iter.takewhile(pred, seq)
}
\arguments{
\item{pred}{iterables}

\item{seq}{iterables}
}
\value{
list or vector
}
\description{
itertools.takewhile
}
\examples{
iter.takewhile(function(x) { x > 2 }, c(1, 2, 3, 2, 1))
}

