% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_export.R
\name{py_export}
\alias{py_export}
\title{py_export}
\usage{
py_export(pkgName, libPath = "default")
}
\arguments{
\item{pkgName}{Name under which functions should be made available}

\item{libPath}{alternative path to load saved pythonic.Rds file from}
}
\value{

}
\description{
py_export
}
\examples{
# Clear global enviroment
rm(list = ls())
# Create a function
t<-function(){
  print("hello world")
}
# Export it to an arbitrary package name
py_export(pkgName = "hello")
# Clear global enviroment
rm(list = ls()) #t is no longer available

# Reload based on package name
py_import(pkgName = "hello")
t()
}
