% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_import.R
\name{py_import}
\alias{py_import}
\title{py_import}
\usage{
py_import(pkgName, global = T, libPath = "default")
}
\arguments{
\item{pkgName}{Name under which functions have been saved}

\item{global}{load each function in the global_enviroment (True) or into a list of function (false)}

\item{libPath}{alternative path to load saved pythonic.Rds file from}
}
\value{

}
\description{
py_import
}
\examples{
# Clear global enviroment
rm(list = ls())
# Create a function
t<-function(){
  print("hello world")
}
# Export it to an arbitrary package name
py_export(pkgName = "hello")
# Clear global enviroment
rm(list = ls()) #t is no longer available

# Reload based on package name
py_import(pkgName = "hello")
t()
}
