% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_jupyter.R
\name{py_jupyter}
\alias{py_jupyter}
\title{py_jupyter}
\usage{
py_jupyter(path = getwd(), conda_env = "spyder-env")
}
\arguments{
\item{path}{get current working directory}

\item{conda_env}{conda enviroment to use}
}
\description{
py_jupyter
}
