% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_list_all.R
\name{py_list_all}
\alias{py_list_all}
\title{py_list_all}
\usage{
py_list_all(libPath = "default", to_clipboard = T, as_code = T, pattern = ".+")
}
\arguments{
\item{libPath}{alternative path to load saved pythonic.Rds file from}

\item{to_clipboard}{Boolean: Save the result to the clipboard?}

\item{as_code}{Boolean: Return the result as executable r code (T) or as a character vector with package names (F)}

\item{pattern}{Searchpattern to filter packages by}
}
\value{

}
\description{
py_list_all
}
